Extracted episodes fragment
