+++ app/src/main/res/menu/episodes.xml
@@ -22,6 +22,7 @@
         android:icon="?attr/ic_filter"
         android:menuCategory="container"
         android:title="@string/filter"
+        android:visible="false"
         custom:showAsAction="ifRoom"/>
 
     <item
